Fix large-corpus index build failures (#431, #432, #433)
Large corpora (tens of millions of terms, hundreds of millions of docs,
multi-GB maintenance_work_mem) broke index build in three independent
ways. Fix all three.

#431 - MaxAllocSize overruns in build/merge:
Corpus- and doc-scaled arrays used plain palloc/palloc0/repalloc and
crossed Postgres' 1 GB MaxAllocSize, erroring out mid-build. Convert the
scaled allocations in the serial and parallel segment writers
(build_context.c), the N-way merge (merge.c), and the merged docmap to
the huge variants (palloc_extended(MCXT_ALLOC_HUGE), repalloc_huge),
matching the huge allocations already used elsewhere on these paths.

#432 - silent 32-bit truncation:
- tp_segment_writer_write() and merge_sink_write() took a uint32 length.
  A skip-index write of skip_entries_count * sizeof(TpSkipEntry) is a
  Size and silently truncated past 4 GiB, producing a corrupt segment.
  Widen the length parameters to Size.
- Three string_offset locals in the segment scan path were uint32 but
  assigned header->strings_offset (uint64) + value, truncating the read
  offset. Widen them to uint64 (tp_segment_read already takes uint64).
- String-pool offsets are genuinely uint32 in the on-disk format. Guard
  the offset computation and raise ERRCODE_PROGRAM_LIMIT_EXCEEDED before
  writing, instead of wrapping silently, when a single segment's string
  pool would exceed the format limit.

#433 - flush budget can exceed the 4 GiB arena:
The in-memory build arena addresses at most ARENA_MAX_PAGES *
ARENA_PAGE_SIZE (4 GiB); ArenaAddr packs page index and offset into 32
bits. The flush budget was set straight from maintenance_work_mem and
never clamped, so a budget at or above 4 GiB meant the spill check never
fired and the arena hit its hard ceiling. Add ARENA_MAX_BYTES and
tp_arena_clamp_budget() (clamp to 3/4 capacity, with headroom for
per-document overshoot) and apply it to the serial (build.c) and
per-worker (build_parallel.c) budgets. Also correct the
TP_MAX_WORKER_SEGMENTS errhint, which advised reducing
max_parallel_maintenance_workers - that makes the problem worse.

+++ src/segment/scan.c
@@ -106,7 +106,7 @@ tp_segment_posting_iterator_init(
 		TpStringEntry string_entry;
 		int			  cmp;
 		uint32		  string_offset_value;
-		uint32		  string_offset;
+		uint64		  string_offset;
 
 		mid = left + (right - left) / 2;
 
@@ -618,7 +618,7 @@ tp_segment_get_doc_freq(
 			TpStringEntry string_entry;
 			int			  cmp;
 			uint32		  string_offset_value;
-			uint32		  string_offset;
+			uint64		  string_offset;
 			int			  mid = left + (right - left) / 2;
 
 			tp_segment_read(
@@ -747,7 +747,7 @@ tp_batch_get_segment_doc_freq(
 			{
 				int	   mid = left + (right - left) / 2;
 				uint32 string_offset_value;
-				uint32 string_offset;
+				uint64 string_offset;
 				uint32 string_length;
 				int	   cmp;
